refactor: move componentes para pasta de componente
