+++ poker-planning-v3/src/frontend/components/ConnectionStatusBanner.tsx
