+++ poker-planning-v3/src/frontend/components/EnterPoker.tsx
@@ -8,9 +8,9 @@ import {
   DialogTitle,
   DialogTrigger,
 } from "@/frontend/components/ui/dialog";
-import { Button } from "./components/ui/button";
-import { Label } from "./components/ui/label";
-import { Input } from "./components/ui/input";
+import { Button } from "./ui/button";
+import { Label } from "./ui/label";
+import { Input } from "./ui/input";
 import { useCallback, useEffect, useState } from "react";
 import { usePokerGameContext } from "./PokerGameContext";
