+++ poker-planning-v3/src/frontend/components/PossibleCards.tsx
@@ -1,4 +1,4 @@
-import { Card } from "./components/ui/Card";
+import { Card } from "./ui/Card";
 import { usePokerGameContext } from "./PokerGameContext";
 import { useCallback, useMemo } from "react";
